Directed at production problem related to dependency shift
ActiveFedora (in our old-ass version) still expects to be able to
`return false` out of a callback to skip remaining callbacks.

This buys us some time to still use rails 5.1 with ActiveFedora, but it
is fleeting, since the option goes away in rails 5.2.

See: https://github.com/sul-dlss/hydrus/issues/161