Turn off publish and shelve when dark in missed spots
This commit patches more code paths that set the publish and shelve attributes, ensuring that dark objects have *no* files set to publish or shelve EVEN IF the stub content metadata says so. This is a problem with some Goobi objects, and this patch allows those items to continue accessioning.