Handle referencing files in druid subdirectory
Rather than requiring the user of the Abbyy class to know that file
paths should be in a druid subdirectory using Windows file separator the
caller can just pass in the filenames and the Abbyy class will handle
making sure they are referenced in the XML using the druid subdirectory.

Even though they are currently out of scope the ticket writing will also
handle when there is file hierarchy in the path: 'x/y.png' instead of
'y.png'.