Create a script that can be used to recover from ABBYY watcher failures
This adds a script that can be run in the event that the watcher process
goes down for some time while OCR is being processed. It checks ABBYY's
result and exception directories and manually updates the workflow
status for any items recorded there, so that the system can be informed
of changes that happened while the watcher was down.

This was tested by putting success/failure result XML files in the
relevant directories locally and invoking the script.

It's intended to not halt when there is an error communicating with
DSA or the workflow service, because some items may already have had
their status updated, so the calls to update them again will fail.