Submit speech-to-text job to AWS Batch
To allow for automatic provisioning of GPU ready EC2 instances we are
going to use AWS Batch instead of running a dedicated EC2 instance. This
means that work needs to be submitted to a Batch Queue instead of SQS.
Since Batch doesn't have any facility for sending notifications on job
completion we are still going to keep the "done" SQS queue.

Resolves #1453